Changed query selectors on the buttons, provided logic to evaluate the operators, and render them on the screen
Instead of querySelectAll on all button elements, I have different id selectors to the numbers and operators. With this, I created three empty arrays to store the first number, the operator, and second number. I added event listeners to each operand and operator when clicked, and made two more arrays that show the result and evaluation process in the display screen on the calculator. A boolean flag was made to check the condition if the operator was clicked or not. I also used css to position the elements in the display screen